feat(cacheLastIds.js): add script to pre-cache last ids of products in Redis
This commit adds a script to pre-cache the last ids of products in Redis. The script queries the database for products in batches of 5 and caches the last id of each batch in Redis. This is done to improve the performance of the application by reducing the number of database queries needed to retrieve the last id of a product.